Internal Link Bug fix
- ID Attributes should no longer have space characters appearing
in the output.
- Internal links should no longer output URL contain space characters,
so that these links can actually jump to anchors that are named with
spaces.
- Added two test cases with an internal link pointing to an identifier
with spaces: The spaces in the identifier, as well as the URL field
of the hyperlink should both be squished. Other types of links are
unaffected with this change.

- Potential problem: An identifier with two words named using
camel case, and another identifier with the same words separated
by spaces are equivalent identifier values. The fix is a necessary
evil for conformance with HTML validation while preserving
functionality.